requirements.txt
fastapi
uvicorn
pymupdf
pydantic
python-multipart
fitz
python-docx
dotenv
requests
streamlit